services/opengl_renderer/requirements.txt
opencv-python
pygame
PyOpenGL==3.1.6
Pillow
pywavefront